Address Copilot review: uint overload for ParsingErrorLength, fix DebugAssertSuppressor
- Add uint overload of SQL.ParsingErrorLength() so the FeatureExtAck
  bounds check passes the raw uint dataLen without truncating to int.
  This prevents misleading negative values in diagnostics when a wire
  length exceeds int.MaxValue.
- Fix DebugAssertSuppressor: add a static lock to serialize access to
  the global Trace.Listeners collection, and clear before restore in
  Dispose to prevent duplicate listeners.
- Update [Collection] comment to explain why serialization is required
  (DebugAssertSuppressor mutates global state).

+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/TdsParser.cs
@@ -3852,7 +3852,7 @@ namespace Microsoft.Data.SqlClient
                     }
                     if (dataLen > (uint)TdsEnums.MaxTokenDataLength)
                     {
-                        throw SQL.ParsingErrorLength(ParsingErrorState.CorruptedTdsStream, (int)dataLen);
+                        throw SQL.ParsingErrorLength(ParsingErrorState.CorruptedTdsStream, dataLen);
                     }
                     int dataLength = (int)dataLen;
                     byte[] data = new byte[dataLength];
